requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
ExifRead==3.0.0
packaging==24.2
pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.1.0
sqlparse==0.5.3
